refactor: LottoController에서 비즈니스 로직을 LottoService으로 분리
