+++ src/controllers/LottoController.js
@@ -1,21 +1,23 @@
-import Lotto from "../models/Lotto.js";
 import InputView from "../views/InputView.js";
 import OutputView from "../views/OutputView.js";
 import InputValidator from "../validators/InputValidator.js";
-import { WINNING_CRITERIA, THOUSAND_UNIT, HUNDRED_PERCENT, MINIMUM_INDEX, PICK_AMOUNT, RANK_INDEX } from "../constants/Constants.js";
+import LottoService from "../services/LottoService.js";
+import { WINNING_CRITERIA, RANK_INDEX, MINIMUM_INDEX } from "../constants/Constants.js";
+
 
 class LottoController {
     constructor() {
         this.InputView = new InputView();
         this.outputView = new OutputView();
+        this.lottoService = new LottoService();
     }
 
     async start() {
         const inputAmount = await this.getValidatedInputAmount();
-        const lottoCount = this.getLottoCount(inputAmount);
+        const lottoCount = this.lottoService.getLottoCount(inputAmount);
         this.outputView.outputLottoCount(lottoCount);
 
-        const lottos = this.getGeneratedLottos(lottoCount);
+        const lottos = this.lottoService.getGeneratedLottos(lottoCount);
         lottos.forEach((lotto) => this.outputView.outputLotto(lotto));
 
         const inputWinningNumbers = await this.getValidatedInputWinningNumbers();
@@ -23,10 +25,10 @@ class LottoController {
 
         this.outputView.outputStatistics();
 
-        let winningRecord = await this.getWinningRecord(lottos, inputWinningNumbers, inputBonusNumber);
+        let winningRecord = await this.lottoService.getWinningRecord(lottos, inputWinningNumbers, inputBonusNumber);
         await this.getResult(WINNING_CRITERIA, winningRecord);
 
-        const profitRate = await this.getProfitRate(winningRecord, inputAmount);
+        const profitRate = await this.lottoService.getProfitRate(winningRecord, inputAmount);
         this.outputView.outputProfitRate(profitRate);
     }
 
@@ -34,15 +36,6 @@ class LottoController {
         const inputAmount = await this.InputView.getInputAmount();
         InputValidator.isValidLottoAmount(inputAmount);
         return inputAmount;
-    }
-
-    getLottoCount(inputAmount) {
-        return inputAmount / THOUSAND_UNIT;
-    }
-
-    getGeneratedLottos(lottoCount) {
-        const lottos = Array.from({ length: lottoCount }, () => new Lotto());
-        return lottos;
     }
 
     async getValidatedInputWinningNumbers() {
@@ -67,28 +60,11 @@ class LottoController {
         }
     }
 
-    async getWinningRecord(lottos, inputWinningNumbers, inputBonusNumber) {
-        let winningRecord = Array(PICK_AMOUNT).fill(0);
-        lottos.forEach((lotto) => {
-            winningRecord[lotto.convertRank(inputWinningNumbers, inputBonusNumber)]++;
-        });
-        return winningRecord;
-    }
-
     async getResult(winningCriteria, winningRecord) {
         for (let i = RANK_INDEX; i >= MINIMUM_INDEX; i--) {
             const { message } = winningCriteria[i];
             this.outputView.outputResult(message, winningRecord[i]);
         }
-    }
-
-    async getProfitRate(winningRecord, inputAmount) {
-        let profitRate = 0;
-        for (let i = RANK_INDEX; i >= MINIMUM_INDEX; i--) {
-            profitRate += winningRecord[i] * WINNING_CRITERIA[i].price;
-        }
-
-        return (profitRate / inputAmount) * HUNDRED_PERCENT;
     }
 }
